+++ Scripts/Maintenance/Measure-PagePreviewPerformance.ps1
@@ -13,12 +13,6 @@
                 Mandatory = $true,
                 ParameterSetName = 'byMSSession'
         )]
-        [string] $ProjectGUID,
-        [Parameter(
-                Position = 1,
-                Mandatory = $true,
-                ParameterSetName = 'byMSSession'
-        )]
         [string[]] $Tags
     )
     begin {
@@ -32,32 +26,40 @@
         Select-MSSession -UseDefaults ($true);
         Enter-MSSession -UseDefaults ($true);
 
-        Enter-MSProject -ProjectGUID ($ProjectGUID) | Out-Null;
+        Enter-MSProject -ProjectGUID (Get-MSSessionProperty -Name ("ProjectGUID")) | Out-Null;
     }
     process {
         Write-Debug -Message ("[ Process => function {0} ]" -f $MyInvocation.MyCommand);
 
-        $ProjectPages = Get-MSPages -Tags $Tags
+        $CategorySearchConfiguration = @()
+        foreach ($Tag in $Tags)
+        {
+            $CategorySearchConfiguration += [pscustomobject][ordered] @{
+                CategoryGuid = $Tag
+                Operator = "eq"
+            }
+        }
+
+        $ProjectPages = Find-MSPages -CategorySearchConfiguration $CategorySearchConfiguration
         $PerformanceResults = @()
 
         foreach ($ProjectPage in $ProjectPages) {
-            $PerformanceResult = New-Object -TypeName PSObject
-            Add-Member -InputObject $PerformanceResult -MemberType NoteProperty -Name Id -Value $ProjectPage.id
-            Add-Member -InputObject $PerformanceResult -MemberType NoteProperty -Name Guid -Value $ProjectPage.GUID
-            Add-Member -InputObject $PerformanceResult -MemberType NoteProperty -Name Headline -Value $ProjectPage.headline
+            Remove-MSPageCache -PageGuids ($ProjectPage.guid) | Out-Null
+            $PreviewTime = (Measure-Command {Get-MSPagePreview -PageGUID $ProjectPage.guid})
+            $PreviewTimeCached = (Measure-Command {Get-MSPagePreview -PageGUID $ProjectPage.guid})
 
-            Remove-MSPageCache -PageGuids ($PerformanceResult.Guid) | Out-Null
-
-            $PreviewTime = (Measure-Command {Get-MSPagePreview -PageGUID $ProjectPage.GUID})
-            $PreviewTimeCached = (Measure-Command {Get-MSPagePreview -PageGUID $ProjectPage.GUID})
-
-            Add-Member -InputObject $PerformanceResult -MemberType NoteProperty -Name Time -Value "$($PreviewTime.Seconds).$($PreviewTime.MilliSeconds)"
-            Add-Member -InputObject $PerformanceResult -MemberType NoteProperty -Name TimeCached -Value "$($PreviewTimeCached.Seconds).$($PreviewTimeCached.MilliSeconds)"
+            $PerformanceResult = [pscustomobject][ordered]@{
+                Id         = $ProjectPage.id;
+                Guid       = $ProjectPage.guid;
+                Headline   = $ProjectPage.headline;
+                Time       = "$($PreviewTime.Seconds).$($PreviewTime.MilliSeconds)";
+                TimeCached = "$($PreviewTimeCached.Seconds).$($PreviewTimeCached.MilliSeconds)";
+            }
 
             $PerformanceResults += $PerformanceResult
         }
 
-        return $PerformanceResults.GetEnumerator() | Sort-Object -Property TimeCached -Descending
+        return $PerformanceResults
     }
     end {
         Write-Debug -Message ("[ Leave => function {0} ]" -f $MyInvocation.MyCommand);